% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerometR.R
\docType{package}
\name{accelerometR}
\alias{accelerometR}
\alias{accelerometR-package}
\title{Accelerometry model fitting and data analysis}
\description{
A small suite of functions for reading, featurizing, analyzing,
   and visualizing accelerometry data. See examples and vignette for more
   information.
}
\seealso{
\code{\link{read_data}}

\code{\link{transform}}

\code{\link{build}}

\code{\link{compute_class_tables}}

\code{\link{predict_accelerometR}}

\code{\link{plot_accelerometR}}
}

