% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acclR_data.R
\docType{data}
\name{acclR_data}
\alias{acclR_data}
\title{Featurized accelerometry data on 16 subjects with state information}
\format{A data frame with 26005 rows and 19 variables. For details on the
   variables, see \code{\link{transform}}.}
\usage{
acclR_data
}
\description{
A dataset containing featurized accelerometry data on 16
   subjects wearing waist accelerometers. Featurization was done by passing
   the raw data through the $\code{\link{transform}} function. State
   data was obtained through analysis of video footage of the subjects
   during the experiment.
}
\keyword{datasets}

