% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Fits a variety of statistical models to accelerometry data}
\usage{
build(data, training_fraction = 0.75, hmm = TRUE, randomForest = TRUE,
  ntree = 1000, maxnodes = 10)
}
\arguments{
\item{data}{A data frame of accelerometry data. It is required to be in the 
format produced by the transform function in this package.}

\item{training_fraction}{The approximate desired fraction of the total number
of subjects to place in the training set. It is only approximate because
the number of subjects who end up in the training set must be an integer.}

\item{hmm, randomForest}{Logical. If \code{TRUE}, a model of the corresponding
type will be fit to the data. If all are \code{FALSE}, then \code{build()}
simply returns an empty list.}

\item{ntree}{A parameter passed to the
\code{\link[randomForest]{randomForest}} function specifying the number of
trees to grow. See \link[randomForest]{randomForest} for details.}

\item{maxnodes}{A parameter passed to the
\code{\link[randomForest]{randomForest}} function specifying the maximum
number of terminal nodes trees in the forest can have. See
\link[randomForest]{randomForest} for details.}
}
\description{
This is a convenience function that takes care of all of the
   many idiosyncracies involved in fitting models from the various
   statistical packages employed by the accelerometR.
   
   The current version of accelerometR only uses mhsmm and randomForest
   models, but future versions will implememnt more.
}
\examples{
   data(acclR_data)
   results <- build(acclR_data)
}
\seealso{
\code{\link[randomForest]{randomForest}},
   \code{\link[mhsmm]{hmmfit}}
}

