% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_hmm_model.R
\name{build_hmm_model}
\alias{build_hmm_model}
\title{Fits a hidden Markov model to accelerometry data.}
\usage{
build_hmm_model(data)
}
\arguments{
\item{data}{A data frame containing accelerometry data from
\code{\link{transform}}.}
}
\value{
A list of class \code{"hmm"} containing the hidden Markov model fit
   to the input data.
}
\description{
\code{\link[mhsmm]{hmmfit}} requires a lot of very
   particular inputs before it will run. \code{\link{build_hmm_model}} 
   assembles all of the necessary inputs given accelerometry data from
   \code{\link{transform_data}} and return the hidden Markov model
   from running \code{\link[mhsmm]{hmmfit}} on those inputs.
}
\keyword{internal}

