% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_training_set_size.R
\name{choose_training_set_size}
\alias{choose_training_set_size}
\title{Chooses how many subjects in your data place into the training set.}
\usage{
choose_training_set_size(num_subjects, training_fraction)
}
\arguments{
\item{num_subjects}{The number of unique subject id's in the accelerometry
data.}

\item{training_fraction}{The approximate desired fraction of the total number
of subjects to place in the training set. It is only approximate because
the number of subjects who end up in the training set must be an integer.}
}
\value{
An integer representing a best-attempt to put the desired fraction of
   subjects into the training set while still leaving at least one subject
   in the test set. If there is only one subject in the input data, that
   subject will be placed in the training set.
}
\description{
A function internal to the accelerometR package designed to help
   split the input dataset into a training set and test set for the purposes
   of estimating the test classification error. It decides how many test
   subjects should go into the training set, and handles edge cases for when
   the number of subjects is very small.
}
\keyword{internal}

