% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_class_tables.R
\name{compute_class_tables}
\alias{compute_class_tables}
\title{Compute classification tables}
\usage{
compute_class_tables(predicted, actual, id)
}
\arguments{
\item{predicted}{A vector of predicted values.}

\item{actual}{A vector of actual values.}

\item{id}{a factor of subject id's.}
}
\value{
A list of classification tables. The structure of the list is as
   follows:
   
   \code{..$frequency} \cr
   \code{.. ..$by_actual} \cr
   \code{.. ..$by_actual_and_id} \cr
   \code{..$rate} \cr
   \code{.. ..$by_actual} \cr
   \code{.. ..$by_actual_and_id}
   
   The tables in \code{..$frequency} give the actual number of observations
   in each cell. The tables in \code{..$rate} give the row-wise fraction of
   observations in each cell. The tables under \code{.. ..$by_actual} are
   2-dimensional tables, with predicted classes on the first dimension and
   actual classes on the second dimension. The tables under
   \code{.. ..$by_actual_and_id} are 3-dimensional tables, with subject id on
   the first dimension, predicted classes on the second dimension, and actual
   classes on the third dimension. The \code{.. ..$by_actual_and_id} tables
   are there so that the user can explore whether the various models perform
   better on some subjects than on others.
}
\description{
Creates a list of classification tables given predicted values,
   actual values, and subject ids.
}
\examples{
   data(acclR_data)
   predictions <- predict_accelerometR(acclR_data)$randomForest_pred
   class_tables <- compute_class_tables(predictions,
   acclR_data$state, acclR_data$id)
}

