% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_accelerometR.R
\name{plot_accelerometR}
\alias{plot_accelerometR}
\title{Plot predited vs. actual states for easy comparison}
\usage{
plot_accelerometR(predicted, actual, main = "Actual vs. predicted state",
  timelabel = "Timesteps (30 second intervals)")
}
\arguments{
\item{predicted}{A factor vector giving the sequence of state predictions
over time.}

\item{actual}{A factor vector giving the sequence of actual states over time.}

\item{main}{A string giving the title of the resulting plot.}

\item{timelabel}{A string giving the x-axis label of the resulting plot.}
}
\description{
Creates a complex plot comparing one set of state predictions to
   the actual states. Care is taken to highlight the points for which the
   predictioned states \emph{differ} from the actual states.
}
\examples{
   data(acclR_data)
   subject1data <- acclR_data[acclR_data$id == 100010, ]
   predicted <-
   predict_accelerometR(subject1data, hmm = NULL)$randomForest_pred
   actual <- subject1data$state
   plot_accelerometR(predicted, actual,
   main = "Actual vs. predicted state, Subject 100010")
}

