% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Performs a series of useful transformations on raw accelerometry data}
\usage{
transform_data(individual_df, epoch = 30)
}
\arguments{
\item{individual_df, }{a list of dataframes containing individual user's data from read_data()}

\item{epoch, }{an integer specified by the user to indicate the number of seconds between each interval}
}
\value{
A data frame containing the featurized data.
}
\description{
Transform raw accelerometry data into features that can be used
   for modeling and predictions. For more information on the features created
   in this version, see 'details'.
}
\details{
The features created in this version are as follows: \cr
   id: a factor containing the unique identifier of each subject \cr
   timestep: We want this as an integer starting at 0, with each time step
     representing a specified interval of time, or 'epoch.' \cr
   x_axis, y_axis, z_axis: the acceleration data in the x, y, and z
     directions, respectively. \cr
   logdetS: the natural log of the determinant of the covariance matrix of
     the raw acceleration data in the x, y, and z directions for each epoch. \cr
   raw_state: A factor variable with levels "sleep", "wake", and "nonwear". \cr
   state: A factor variable obtained by smoothing raw_state to exclude 
     unlikely events like the subject being asleep for 3 minutes in the
     in the middle of two long stretches of being awake. \cr
   L2_norm: The euclidean norm of the acceleration data in an epoch:
     \code{sqrt(x_axis^2 + y_axis^2 + z_axis^2)}. \cr
   mSm_norm1: Where \code{m} is a vector of (x_axis, y_axis, z_axis)' and
     \code{S} is the covariance matrix of the raw x, y, and z accelerations,
     this is computed by \eqn{m' S m}. \cr
   mSm_norm2: Similar, but for \eqn{m' S^{-1} m}. This is often known as the
     Mahalanobis distance \cr
   mDiff: \eqn{(m_t - m_{t-1})' S_{t-1} (m_t - m_{t-1} )} \cr
   tDiff: \eqn{(m_t - m_{t-1} )' S_{t-1}^{-1} (m_t - m_{t-1} )} \cr
   logSDiff: \eqn{\mbox{logdet}S_{t-1} - \mbox{logdet}S_t}
   theta_x, theta_y, theta_z: \eqn{\arccos}(*_axis / L2norm). These three
     measure the orientation of the device relative to gravity. \cr
   logdetT: the natural log of the determinant of the covariance matrix
     between the angles theta_x, theta_y, and theta_z computed within each
     epoch. \cr
   logTDiff: \eqn{\mbox{logdet}T_{t-1} - \mbox{logdet}T_t}
}
\examples{
   filename <- system.file("extdata", "subject100010.csv",
                         package="accelerometR")
   individual_df <- read_data(path = filename,
                   id = '100010',
                   x = 'waistrawaxis1',
                   y = 'waistrawaxis2',
                   z = 'waistrawaxis3',
                   datetime = 'dttm',
                   rawstate = 'se_event')
   featurized_df <- transform_data(individual_df, epoch = 30)
}

